+++ private/controllers/Schools.php
@@ -17,9 +17,13 @@ class Schools extends Controller{
         $crumbs[] = ['Dashboard',''];
         $crumbs[] = ['Schools','schools'];
 
+        if(Auth::access('super_admin')){ 
         $this->view('schools', 
         ['rows' => $data,
         'crumbs' => $crumbs]);
+    }else{
+        $this->view('access-denied');
+    }
     }
 
     public function add(){
@@ -30,7 +34,7 @@ class Schools extends Controller{
 
         $errors = [];
 
-        if(count($_POST) > 0){
+        if(count($_POST) > 0 && Auth::access('super_admin')){
             $school = new School();
             if($school->validate($_POST)){
 
@@ -46,9 +50,14 @@ class Schools extends Controller{
         $crumbs[] = ['Schools','schools'];
         $crumbs[] = ['Add','schools/add'];
 
-        $this->view('schools.add', [
-            'errors' => $errors,
-            'crumbs' => $crumbs]);
+        if(Auth::access('super_admin')){
+            $this->view('schools.add', [
+                'errors' => $errors,
+                'crumbs' => $crumbs]);
+        }else{
+            $this->view('access-denied');
+        }
+        
     }
 
     public function edit($id = null){
@@ -59,7 +68,7 @@ class Schools extends Controller{
         $school = new School();
         $errors = [];
 
-        if(count($_POST) > 0){
+        if(count($_POST) > 0 && Auth::access('super_admin')){
            
             if($school->validate($_POST)){
 
@@ -76,11 +85,15 @@ class Schools extends Controller{
        $crumbs[] = ['Schools','schools'];
        $crumbs[] = ['Edit','schools/edit'];
 
+       if(Auth::access('super_admin')){
         $this->view('schools.edit', [
             'row' => $row,
             'errors' => $errors,
             'crumbs' => $crumbs
         ]);
+    }else{
+        $this->view('access-denied');
+    }
     }
 
     public function delete($id = null){
@@ -91,7 +104,7 @@ class Schools extends Controller{
         $school = new School();
         $errors = [];
 
-        if(count($_POST) > 0){
+        if(count($_POST) > 0 && Auth::access('super_admin')){
            
 
                 $school->delete($id);
@@ -105,9 +118,13 @@ class Schools extends Controller{
        $crumbs[] = ['Schools','schools'];
        $crumbs[] = ['Delete','schools/delete'];
        
+       if(Auth::access('super_admin')){
         $this->view('schools.delete', [
             'row' => $row,
             'crumbs' => $crumbs
         ]);
+    }else{
+        $this->view('access-denied');
+    }
     }
 }
